update feature added

+++ shop/products/forms.py
@@ -1,11 +1,11 @@
 from flask_wtf import FlaskForm
 from flask_wtf.file import FileAllowed, FileField, FileRequired
-from wtforms import StringField, SubmitField, IntegerField, TextAreaField, validators
+from wtforms import StringField, SubmitField, IntegerField, TextAreaField, validators, DecimalField
 
 
 class Addproducts(FlaskForm):
     name = StringField('Name',[validators.DataRequired()])
-    price = IntegerField('Price',[validators.DataRequired()])
+    price = DecimalField('Price',[validators.DataRequired()])
     discount = IntegerField('Discount', default = 0)
     stock = IntegerField('Stock',[validators.DataRequired()])
     description = TextAreaField('Description', [validators.DataRequired()])
